feat(core): throw an error if size is greater than 2MB
